Consolidate ListingForm component location
